orchestrator/requirements.txt
fastapi
uvicorn
python-multipart
pydantic
pytest
httpx
pytest-mock
transformers
torch
redis
docker
GitPython
